refactor: 경로 index.js 생략
