+++ src/io/InputHandler.js
@@ -1,7 +1,7 @@
 import { Console } from '@woowacourse/mission-utils';
-import { InputValidator } from '../services/index.js';
+import { InputValidator } from '../services';
 import { tryAgain } from '../utils/validateUtils.js';
-import { CONFIG } from '../constants/index.js';
+import { CONFIG } from '../constants';
 
 class InputHandler {
   async processMoneyInput(query) {
