+++ src/io/Printer.js
@@ -1,5 +1,6 @@
+// 사용 시 Printer 네임스페이스로 import
 import { Console } from '@woowacourse/mission-utils';
-import { CONFIG } from '../constants/index.js';
+import { CONFIG } from '../constants';
 
 export const print = (query) => {
   Console.print(query);
